Refactor view helpers
