+++ app/views/helpers.py
@@ -1,0 +1,211 @@
+import asyncio
+from contextlib import suppress
+
+from sanic import exceptions, response
+from sanic.log import logger
+
+from .. import models, settings, utils
+
+
+async def generate_url(
+    request, template_id: str = "", *, template_id_required: bool = False
+):
+    if request.form:
+        payload = dict(request.form)
+        for key in list(payload.keys()):
+            if "lines" not in key and "style" not in key:
+                payload[key] = payload.pop(key)[0]
+    else:
+        try:
+            payload = request.json or {}
+        except exceptions.InvalidUsage:
+            payload = {}
+
+    with suppress(KeyError):
+        payload["style"] = payload.pop("style[]")
+    with suppress(KeyError):
+        payload["text_lines"] = payload.pop("text_lines[]")
+
+    if template_id_required:
+        try:
+            template_id = payload["template_id"]
+        except KeyError:
+            return response.json({"error": '"template_id" is required'}, status=400)
+        else:
+            template_id = utils.text.slugify(template_id)
+
+    style: str = utils.urls.arg(payload, "", "style", "overlay", "alt")
+    if isinstance(style, list):
+        style = ",".join([(s.strip() or "default") for s in style])
+    while style.endswith(",default"):
+        style = style.removesuffix(",default")
+    text_lines = utils.urls.arg(payload, [], "text_lines")
+    font = utils.urls.arg(payload, "", "font")
+    background = utils.urls.arg(payload, "", "background", "image_url")
+    extension = utils.urls.arg(payload, "", "extension")
+
+    if style == "animated":
+        extension = "gif"
+        style = ""
+
+    status = 201
+
+    if template_id:
+        template: models.Template = models.Template.objects.get_or_create(template_id)
+        url = template.build_custom_url(
+            request,
+            text_lines,
+            style=style,
+            font=font,
+            extension=extension,
+        )
+        if not template.valid:
+            status = 404
+            template.delete()
+    else:
+        template = models.Template("_custom")
+        url = template.build_custom_url(
+            request,
+            text_lines,
+            background=background,
+            style=style,
+            font=font,
+            extension=extension,
+        )
+
+    url, _updated = await utils.meta.tokenize(request, url)
+
+    if payload.get("redirect", False):
+        return response.redirect(utils.urls.add(url, status="201"))
+
+    return response.json({"url": url}, status=status)
+
+
+async def preview_image(request, id: str, lines: list[str], style: str):
+    error = ""
+
+    id = utils.urls.clean(id)
+    if utils.urls.schema(id):
+        template = await models.Template.create(id)
+        if not template.image.exists():
+            logger.error(f"Unable to download image URL: {id}")
+            template = models.Template.objects.get("_error")
+            error = "Invalid Background"
+    else:
+        template = models.Template.objects.get_or_none(id)
+        if not template:
+            logger.error(f"No such template: {id}")
+            template = models.Template.objects.get("_error")
+            error = "Unknown Template"
+
+    if not any(line.strip() for line in lines):
+        lines = template.example
+
+    if not utils.urls.schema(style):
+        style = style.strip().lower()
+    if not await template.check(style):
+        error = "Invalid Overlay"
+
+    data, content_type = await asyncio.to_thread(
+        utils.images.preview, template, lines, style=style, watermark=error.upper()
+    )
+    return response.raw(data, content_type=content_type)
+
+
+async def render_image(
+    request,
+    id: str,
+    slug: str = "",
+    watermark: str = "",
+    extension: str = settings.DEFAULT_EXTENSION,
+):
+    lines = utils.text.decode(slug)
+    asyncio.create_task(utils.meta.track(request, lines))
+
+    status = int(utils.urls.arg(request.args, "200", "status"))
+
+    if any(len(part.encode()) > 200 for part in slug.split("/")):
+        logger.error(f"Slug too long: {slug}")
+        slug = slug[:50] + "..."
+        lines = utils.text.decode(slug)
+        template = models.Template.objects.get("_error")
+        style = settings.DEFAULT_STYLE
+        status = 414
+
+    elif id == "custom":
+        url = utils.urls.arg(request.args, None, "background", "alt")
+        if url:
+            template = await models.Template.create(url)
+            if not template.image.exists():
+                logger.error(f"Unable to download image URL: {url}")
+                template = models.Template.objects.get("_error")
+                if url != settings.PLACEHOLDER:
+                    status = 415
+
+            style = utils.urls.arg(request.args, settings.DEFAULT_STYLE, "style")
+            if not utils.urls.schema(style):
+                style = style.lower()
+            if not await template.check(style):
+                if utils.urls.schema(style):
+                    status = 415
+                elif style != settings.PLACEHOLDER:
+                    status = 422
+
+        else:
+            logger.error("No image URL specified for custom template")
+            template = models.Template.objects.get("_error")
+            style = settings.DEFAULT_STYLE
+            status = 422
+
+    else:
+        template = models.Template.objects.get_or_none(id)
+        if not template or not template.image.exists():
+            logger.error(f"No such template: {id}")
+            template = models.Template.objects.get("_error")
+            if id != settings.PLACEHOLDER:
+                status = 404
+
+        style = utils.urls.arg(request.args, settings.DEFAULT_STYLE, "style", "alt")
+        if not await template.check(style):
+            if utils.urls.schema(style):
+                status = 415
+            elif style != settings.PLACEHOLDER:
+                status = 422
+
+    if extension not in settings.ALLOWED_EXTENSIONS:
+        extension = settings.DEFAULT_EXTENSION
+        status = 422
+
+    font_name = utils.urls.arg(request.args, "", "font")
+    if font_name == settings.PLACEHOLDER:
+        font_name = ""
+    else:
+        try:
+            models.Font.objects.get(font_name)
+        except ValueError:
+            font_name = ""
+            status = 422
+
+    try:
+        size = int(request.args.get("width", 0)), int(request.args.get("height", 0))
+        if 0 < size[0] < 10 or 0 < size[1] < 10:
+            raise ValueError(f"dimensions are too small: {size}")
+    except ValueError as e:
+        logger.error(f"Invalid size: {e}")
+        size = 0, 0
+        status = 422
+
+    frames = int(request.args.get("frames", 0))
+
+    path = await asyncio.to_thread(
+        utils.images.save,
+        template,
+        lines,
+        watermark,
+        font_name=font_name,
+        extension=extension,
+        style=style,
+        size=size,
+        maximum_frames=frames,
+    )
+    return await response.file(path, status)
