+++ app/views/templates.py
@@ -1,5 +1,4 @@
 import asyncio
-from contextlib import suppress
 from dataclasses import dataclass
 
 from sanic import Blueprint, exceptions, response
@@ -7,6 +6,7 @@ from sanic_ext import openapi
 
 from .. import helpers, settings, utils
 from ..models import Template
+from .helpers import generate_url
 
 blueprint = Blueprint("Templates", url_prefix="/templates")
 
@@ -121,77 +121,3 @@ class CustomRequest:
 )
 async def custom(request):
     return await generate_url(request)
-
-
-async def generate_url(
-    request, template_id: str = "", *, template_id_required: bool = False
-):
-    if request.form:
-        payload = dict(request.form)
-        for key in list(payload.keys()):
-            if "lines" not in key and "style" not in key:
-                payload[key] = payload.pop(key)[0]
-    else:
-        try:
-            payload = request.json or {}
-        except exceptions.InvalidUsage:
-            payload = {}
-
-    with suppress(KeyError):
-        payload["style"] = payload.pop("style[]")
-    with suppress(KeyError):
-        payload["text_lines"] = payload.pop("text_lines[]")
-
-    if template_id_required:
-        try:
-            template_id = payload["template_id"]
-        except KeyError:
-            return response.json({"error": '"template_id" is required'}, status=400)
-        else:
-            template_id = utils.text.slugify(template_id)
-
-    style: str = utils.urls.arg(payload, "", "style", "overlay", "alt")
-    if isinstance(style, list):
-        style = ",".join([(s.strip() or "default") for s in style])
-    while style.endswith(",default"):
-        style = style.removesuffix(",default")
-    text_lines = utils.urls.arg(payload, [], "text_lines")
-    font = utils.urls.arg(payload, "", "font")
-    background = utils.urls.arg(payload, "", "background", "image_url")
-    extension = utils.urls.arg(payload, "", "extension")
-
-    if style == "animated":
-        extension = "gif"
-        style = ""
-
-    status = 201
-
-    if template_id:
-        template: Template = Template.objects.get_or_create(template_id)
-        url = template.build_custom_url(
-            request,
-            text_lines,
-            style=style,
-            font=font,
-            extension=extension,
-        )
-        if not template.valid:
-            status = 404
-            template.delete()
-    else:
-        template = Template("_custom")
-        url = template.build_custom_url(
-            request,
-            text_lines,
-            background=background,
-            style=style,
-            font=font,
-            extension=extension,
-        )
-
-    url, _updated = await utils.meta.tokenize(request, url)
-
-    if payload.get("redirect", False):
-        return response.redirect(utils.urls.add(url, status="201"))
-
-    return response.json({"url": url}, status=status)
